# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: InvoiceFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products/services
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product/service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product/service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product/service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Invoices
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, sent, paid, overdue]
        - name: search
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvoiceListItem"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a new invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceBody"
      responses:
        "200":
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Invoice deleted
        "404":
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invoices/{id}/mark-paid:
    patch:
      operationId: markInvoicePaid
      tags: [invoices]
      summary: Mark an invoice as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice marked as paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Dashboard
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        defaultRate:
          type: number
        unit:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - defaultRate
        - unit
        - createdAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        defaultRate:
          type: number
        unit:
          type: string
      required:
        - name
        - defaultRate
        - unit
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        defaultRate:
          type: number
        unit:
          type: string
    InvoiceItem:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        productId:
          type: ["integer", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        rate:
          type: number
        amount:
          type: number
      required:
        - id
        - invoiceId
        - name
        - quantity
        - unit
        - rate
        - amount
    CreateInvoiceItemBody:
      type: object
      properties:
        productId:
          type: integer
        name:
          type: string
        description:
          type: string
        quantity:
          type: number
        unit:
          type: string
        rate:
          type: number
        amount:
          type: number
      required:
        - name
        - quantity
        - unit
        - rate
        - amount
    Invoice:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        customerAddress:
          type: ["string", "null"]
        customerCity:
          type: ["string", "null"]
        customerState:
          type: ["string", "null"]
        customerGstin:
          type: ["string", "null"]
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
        dueDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        subtotal:
          type: number
        discountType:
          type: string
          enum: [percent, flat]
        discountValue:
          type: number
        taxPercent:
          type: number
        taxAmount:
          type: number
        total:
          type: number
        notes:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        showBankDetails:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceNumber
        - invoiceDate
        - status
        - subtotal
        - discountType
        - discountValue
        - taxPercent
        - taxAmount
        - total
        - showBankDetails
        - items
        - createdAt
        - updatedAt
    InvoiceListItem:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
        dueDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        total:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceNumber
        - invoiceDate
        - status
        - total
        - createdAt
    CreateInvoiceBody:
      type: object
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        customerCity:
          type: string
        customerState:
          type: string
        customerGstin:
          type: string
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
        dueDate:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        discountType:
          type: string
          enum: [percent, flat]
        discountValue:
          type: number
        taxPercent:
          type: number
        notes:
          type: string
        paymentTerms:
          type: string
        showBankDetails:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateInvoiceItemBody"
      required:
        - invoiceNumber
        - invoiceDate
        - status
        - items
    UpdateInvoiceBody:
      type: object
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        customerCity:
          type: string
        customerState:
          type: string
        customerGstin:
          type: string
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
        dueDate:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        discountType:
          type: string
          enum: [percent, flat]
        discountValue:
          type: number
        taxPercent:
          type: number
        notes:
          type: string
        paymentTerms:
          type: string
        showBankDetails:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateInvoiceItemBody"
...